vistas

+++ app/Http/Controllers/CapituloController.php
@@ -22,6 +22,8 @@ class CapituloController extends Controller
      */
     public function create()
     {
+        $this->authorize('administrar');
+
         return view('capitulos.create');
     }
 
@@ -30,6 +32,8 @@ class CapituloController extends Controller
      */
     public function store(StoreCapituloRequest $request)
     {
+        $this->authorize('administrar');
+
         Capitulo::create($request->validated());
         return redirect()->route('capitulos.index');
     }
@@ -47,6 +51,8 @@ class CapituloController extends Controller
      */
     public function edit(Capitulo $capitulo)
     {
+        $this->authorize('administrar');
+
         return view('capitulos.edit', compact('capitulo'));
     }
 
@@ -55,6 +61,8 @@ class CapituloController extends Controller
      */
     public function update(UpdateCapituloRequest $request, Capitulo $capitulo)
     {
+        $this->authorize('administrar');
+
         $capitulo->update($request->validated());
         return redirect()->route('capitulos.index');
     }
@@ -64,6 +72,8 @@ class CapituloController extends Controller
      */
     public function destroy(Capitulo $capitulo)
     {
+        $this->authorize('administrar');
+        
         $capitulo->delete();
         return redirect()->route('capitulos.index');
     }
